pyproject.toml
[project]
name = "aws_infrastructure_security_audit_and_reporting"
version = "0.1.0"
description = "AWS Infrastructure Security Audit and Reporting using CrewAI"
authors = [
    {name = "Tony Kipkemboi", email = "tony@crewai.com"},
]
dependencies = [
    "crewai[tools]>=0.16.0",
    "boto3>=1.34.0",
    "python-dotenv>=1.0.0",
    "ollama>=0.1.7",
    "langchain-ollama>=0.1.0",
    "llama-cpp-python>=0.2.0",
    "langchain-community>=0.2.0",
]
requires-python = ">=3.10,<3.14"
readme = "README.md"
license = {text = "MIT"}

[project.scripts]
aws_infrastructure_security_audit_and_reporting = "aws_infrastructure_security_audit_and_reporting.main:run"
test = "aws_infrastructure_security_audit_and_reporting.main:test"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
